Refactor auth module interfaces and remove user interface
