+++ src/modules/auth/interfaces/index.ts
@@ -1,4 +1,3 @@
 export * from './auth.response'
 export * from './auth.status'
 export * from './login.interfaces'
-export * from './user.interface'
